requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
python-decouple>=3.8
dj-rest-auth>=3.0
Pillow>=10.0
drf-yasg>=1.21